# GeDeCo/GeDeCo/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/equipos/registrar":{"post":{"tags":["equipo-controller"],"operationId":"registrarEquipoEnCompeticion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EquipoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/competiciones/registrar":{"post":{"tags":["competicion-controller"],"operationId":"registrarCompeticion","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompeticionDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/competiciones/jornada/{competicionID}":{"get":{"tags":["competicion-controller"],"operationId":"consultarPartidosPrimeraJornada","parameters":[{"name":"competicionID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["competicion-controller"],"operationId":"generarPrimeraJornada","parameters":[{"name":"competicionID","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/equipos/listarTodos":{"get":{"tags":["equipo-controller"],"operationId":"listarTodosLosEquipos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/equipos/consultarPorCompeticion/{idCompeticion}":{"get":{"tags":["equipo-controller"],"operationId":"consultarEquiposPorCompeticion","parameters":[{"name":"idCompeticion","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/competiciones/listarTodas":{"get":{"tags":["competicion-controller"],"operationId":"listarTodasLasCompeticiones","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"EquipoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"competicionID":{"type":"integer","format":"int64"}}},"CompeticionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"deporte":{"type":"string"},"fechaInicio":{"type":"string","format":"date"},"fechaFin":{"type":"string","format":"date"},"numPistas":{"type":"integer","format":"int32"},"diasReservados":{"type":"string"}}}}}}